requirements.txt
langchain
langgraph
langchain-google-genai
python-dotenv
langchain-ollama
fastapi
uvicorn
